requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
requests==2.32.3

SQLAlchemy==2.0.34
PyMySQL==1.1.1

Jinja2==3.1.4
python-multipart==0.0.9

APScheduler==3.10.4